Create README.md
This MATLAB code and the accompanying MATLAB app installer can be used to create structured mesh in a trapezoidal geometry. The app also allows the 
extraction of the coordinates of the nodes of the mesh and the connectivity table which can be used in the computation involving finite element 
method.